Update Promise, PromiseConverter for better error propagation
Previously, the PromiseConverter would always pass promise failures on
to the TaskSet for central error handling. If QML handled the error in a
context-specific way, central error handling also received the error.
Now, the error is only propagated to the TaskSet for central handling if
QML does not set a handler.