Make over-long contestant descriptions fade at bottom
Now if a contestant description is too long, instead of eliding it, we
fade it out at the bottom. This is a slightly clearer indication to the
user that more text is available, but hidden. As an added bonus, this
fixes an issue on my computer where the app would hang when scrolling a
contest list.