Progress #111: Add creator signature field to ContestCreationRequest
Wrap the public contest details in a group so they're easier to
serialize and sign, and add a creator signature field to the request.

Ultimately, the contestOptions group will probably be dropped and a
reference to a structure in contest.capnp will be used instead, but
that's a bigger refactor than I have time for just this moment.