Cleanup: Make contest datagram key a well-defined type
Previously, the key on the contest datagram as a Data field containing an
FC-serialized std::pair. This is super dirty, so I've changed it to
contain a capnp-serialized struct, which is defined in datagram.capnp.
This is much cleaner.

Also, a minor cleanup in CustomEvaluator: create BlobMessageReader,
which makes creating a MessageReader from a Data::Reader a single-line
affair.