Initial work on config for GrapheneBackend
Graphene's config system is sufficient for basic purposes, but eventually
we'll want a much more advanced config system built into the Graphene
Backend. This commit lays some groundwork for that system.

In particular, I want a config system that's persistent (no duh!) but can
be changed live at runtime without interruptions to service. This means
that the config will be running live in RAM, but a client can connect via
some API and send in a new config that will immediately take effect
(replacing the live one), and then be saved to disk to persist it for
future runs. CapnProto works nicely for this if we use the `capnp eval`
form of config files as described at the bottom of this page: https://
capnproto.org/capnp-tool.html