Resolve #33: UNTESTED: Finish contest search implementation
Early implementation of contest search. There's a fair bit of premature
optimization done here, but not as much as I thought about doing. :)
This is a very rough search system, which can be greatly improved in the
future (in terms of both optimization and functionality). This is just
enough to get the job done.