Resolve #126: Replace and remove BlockchainAdaptorInterface
I'm really glad I did this; there's a lot of really sketchy voodoo
that's gone now in how data is managed and keeping track of ownership.
Admittedly, it's a bit more complex now since the VotingApp actually has
to keep track of data lifespans now, but that's really how it should've
been to begin with. The only reason things worked before is because of
the way my dummy data was stored.