[Refactor] Use structs for IDs rather than Data blobs
The IDs on objects are opaque to all but the blockchain integration
code. To reflect this, I was having the app handle them as Data blobs,
and the chain integration code would serialize/deserialize these blobs
as necessary. This was becoming increasingly toxic, though, as more and
more such opaque fields were introduced, and there was no clear standard
for how they should be formatted.

This commit defines structs for the ID types, and replaces (hopefully)
all of the opaque blobs with actual structs. This makes the code much
easier to work with, and it makes IDs type-safe as the compiler now
knows their types. This should eliminate a whole class of bugs that was
probably growing as I kept changing the contents of the ID types. :)

The notable exception is that QML still treats IDs as opaque hexadecimal
dumps, since QML should never need to parse an ID. The conversions from
structs to/from hex dumps happens within the DataStructures wrappers,
where it should be easy to verify that the types are being handled
correctly.