Streamline startup process
Add and implement an unlockWallet call to wallet API. This allows me to
open the unlock wallet prompt as soon as I connect to the wallet, when I
know I'm going to need it unlocked in just a moment, which allows the
slow human to type his password while the app waits on other calls and
initializations.

In short, parallelize startup operations to accelerate coding
iterations. Also, test out JSON-RPC 2.0 notifications.