Fix server crash when processing decision
If the server processed a decision while the client was subscribed to
results on the contest (which happens all the time) the server would at
least throw an exception and fail to notify the client, and often crash.

This was because the server was failing to fill out the notification of
new results correctly (did not allocate space a list before populating
it). This is now fixed.