Resolve #148
Fixed in two different ways!

First was a bug: if the results object on the ContestResultsApi was set
to nullptr, the change was ignored leaving the API holding a dangling
pointer (can we just pretend I'm not the one who wrote that code?)

Second is a minor optimization: go ahead and delete the
ContestResultsApi when the ContestDetailPage is closed rather than
waiting for the garbage collector to find it. This releases server
resources sooner.