Progress #152: Change actions based on status
Add a 'Revoke Vote' button, and show it when the contest is in Official
Vote status. Show the Cancel/Cast buttons when the contest is in Pending
Vote status. Hide the status row altogether when the contest has no
official vote and no pending vote.

And make it all pretty =D