Resolve #191
The zero stake issue was because, for that particular voter, the wallet
was reporting the balance as a JSON number rather than a string, and
apparently Qt does not correctly turn a number into a string when using
toString() on it. Fixed by adding a check: if the weight comes as a
number, parse it as such; otherwise, parse it as a string and then parse
the string to a number.